Add Tracking Footer

Add Tracking Footer

ggtrack(
  gg,
  qr_content,
  color = "black",
  caption = NULL,
  logo = NULL,
  order = "CLQ",
  positions = c(55, 25, 20),
  logo_justification = 1,
  qr_justification = 1,
  height_plot = 7,
  height_tracker = 1.8,
  interactive = FALSE,
  plotly_heights = c(0.8, 0.2),
  ...
)

Arguments

gg
ggplot object to track

qr_content
character content passed to qrencode. A time stamp is automatically added plus the current git commit where available.

color
character color of the QR code

caption
character or grob to add to footer. Text can be html or md and is passed directly to richtext_grob

logo
file to add to footer as a logo

logo_justification
numeric between 0 and 1 passed to rasterGrob. See note below.

qr_justification
numeric between 0 and 1 passed to rasterGrob. See note below.

height_plot
integer centimeters

height_tracker
integer centimeters

interactive
logical use plotly for interactivity

plotly_heights
numeric vector of length 2 to fix relative height of plot and tracking banner

...
options passed to theme in order to style the tracker banner.

Value

Note

For Justification you need to imagine each element in its own rectangle with a bottom dimension of 0 to 1. If you want the logo or QR code on the right of the rectangle set justification to 1, or 0 for left.