Tracker Download Link

Save tracker object to file and render link or button for user download

make_download(
  tracker,
  save_file,
  download_file,
  ext = "png",
  link_text = "download",
  type = c("link", "button"),
  height = 13,
  height_units = c("cm", "mm", "in"),
  button_style = htmltools::css(background.color = "DodgerBlue", border = "none", color
    = "white", padding = "12px 30px", cursor = "pointer", font.size = "16px", width =
    "150px"),
  date = format(Sys.time(), "%Y%m%d-%H%M%S"),
  render = TRUE
)

Arguments

tracker
tracker object to be saved

save_file
character vector, where length > 1, path is generated using file.path

download_file
character vector, where length > 1, path is generated using file.path

link_text
character text to appear on the page

type
character must be one of "link" or "button".

height
integer height value passed to ggsave

height_units
character units value passed to ggsave

button_style
css object generated by css or a css character

render
Boolean, render in-place or save for later

Value

html of the link object